임시 페이지로 구글 로그인 테스트 및 게시글 CRUD 테스트

+++ src/main/java/com/bong/DailyShareService/domain/posts/PostsRepository.java
@@ -13,5 +13,5 @@ public interface PostsRepository extends JpaRepository<Posts, Long> {
 
     @Modifying
     @Query("update Posts p set p.viewCount = p.viewCount + 1 where p.id = :id")
-    Long increaseViewCount(Long id);
+    int increaseViewCount(Long id);
 }
